🚑 Quickfix(offset_encodings): address nvim-nightly offset_encoding deprecations
This is a temporary hotfix to deal with offset_encoding/position_encoding requirements in the latest
nightly builds. Plugins should be updated to address this.

BREAKING CHANGE: blink.cmp is not working, fallback to nvim-cmp. noice.nvim using a fork. Likely
more to come. Watch for 'offset_encoding' scope tags in later commmits.